Move FollowTarget Component Into Separate Files
- Created source and include directories for components.
- Moved the FollowTarget Component from Spectacle.cpp into separate
header and source files, and renamed it to LinearFollowTarget.
